+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Change Log
 
-#### [2.3.0(16)]
+#### [2.3.0(16)](https://github.com/VakhoKontridze/VCore/releases/download/2.3.0/VCore.xcframework.zip) — *2022 02 25*
 
 - `ScrollableView` has been added
 - `animated` and `completion` parameters can now be passed to methods in `StandardNavigable`